Tests pass
Probably not munging properly. Should extract the munging funcs
to utils. The original code to munge only ran once at the start
but now we don't know amount at that point with soft_entry so we
need to do the munging checks later.

If we move the munging funcs to utils we can have the EE fall
over if it gets passed unmunged values.